% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.simss_mielke.R
\name{as.numeric.simss_mielke}
\alias{as.numeric.simss_mielke}
\title{Coerce a Mielke result to a numeric sample size}
\usage{
\method{as.numeric}{simss_mielke}(x, ...)
}
\arguments{
\item{x}{A \code{simss_mielke} object.}

\item{...}{Unused additional arguments.}
}
\description{
Coerce a Mielke result to a numeric sample size
}
