% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calopt.R
\name{calopt}
\alias{calopt}
\title{Sample Size Calculation for Target Power}
\usage{
calopt(
  mu_list,
  varcov_list = NA,
  sigma_list = NA,
  cor_mat = NA,
  sigmaB = NA,
  Eper = c(0, 0),
  Eco = c(0, 0),
  rho = 0,
  TAR = NA,
  arm_names = NA,
  ynames_list = NA,
  type_y = NA,
  list_comparator = NA,
  list_y_comparator = NA,
  power = 0.8,
  alpha = 0.05,
  lequi.tol = NA,
  uequi.tol = NA,
  list_lequi.tol = NA,
  list_uequi.tol = NA,
  dtype = "parallel",
  ctype = "ROM",
  vareq = T,
  lognorm = T,
  k = NA,
  adjust = "no",
  dropout = NA,
  nsim = 5000,
  lower = 2,
  upper = 500,
  seed = 1234,
  ncores = NA,
  step.power = 6,
  step.up = TRUE,
  pos.side = FALSE,
  maxiter = 1000
)
}
\arguments{
\item{mu_list}{Named list of arithmetic means per treatment arm. Each element contains a vector (i.e., one per treatment arm) with the expected outcomes for all endpoints of interest.}

\item{varcov_list}{list of var-cov matrices, each element corresponds to a comparator with a varcov matrix of size number of endpoints X number of endpoints.}

\item{sigma_list}{list of sigma vectors, each element corresponds to a comparator with a sigma vector of size number of endpoints.}

\item{cor_mat}{matrix specifying the correlation matrix between endpoints, used along with sigma_list  to calculate the varcov list in case it is not provided.}

\item{sigmaB}{number between subject variance only for 2x2 design.}

\item{Eper}{vector of size 2, effect of period on c(0,1).}

\item{Eco}{vector of size 2, carry over effect of arm c(Reference, Treatment).}

\item{rho}{correlation parameter to fix on all the pair of endpoints, used along with sigma_list to calculate the varcov list in case neither cor_mat or var-cov list is not provided.}

\item{TAR}{vector of allocation rates with allocation rates of the arm, default is equivalent rate.}

\item{arm_names}{Optional vector with the treatment names. If not supplied, it will be derived from mu_list.}

\item{ynames_list}{Optional list of vectors with Endpoint names on each arm. When not all endpoint names are provided for each arm, arbitrary names (assigned by vector order) are used.}

\item{type_y}{vector with the type of endpoints: primary endpoint(1), otherwise (2).}

\item{list_comparator}{list of comparators, i.e each comparator is a vector of size 1 X 2 where are specified the name of treatments}

\item{list_y_comparator}{list of endpoints to be considered in each comparator. Each element of the list is a vector containing the names of the endpoints to compare. When it is not provided, all endpoints present in both compared arms are used.}

\item{power}{target power (default = 0.8)}

\item{alpha}{alpha level (default = 0.05)}

\item{lequi.tol}{lower equivalence bounds (e.g., -0.5) expressed in raw scale units (e.g., scalepoints) of endpoint repeated on all endpoints and comparators}

\item{uequi.tol}{upper equivalence bounds (e.g., -0.5) expressed in raw scale units (e.g., scalepoints) of endpoint repeated on all endpoints and comparators}

\item{list_lequi.tol}{list of lower equivalence bounds (e.g., -0.5) expressed in raw scale units (e.g., scalepoints) of endpoint in comparator}

\item{list_uequi.tol}{list of upper equivalence bounds (e.g., -0.5) expressed in raw scale units (e.g., scalepoints) of endpoint in comparator}

\item{dtype}{design type ("parallel","2x2")}

\item{ctype}{comparison dtype ("DOM"(Difference of means), "ROM"(Ratio of means))}

\item{vareq}{variance equivalence assumption (FALSE, TRUE)}

\item{lognorm}{Is data log-normally distributed? (TRUE, FALSE)}

\item{k}{Vector with the number of endpoints that must be successful (integer) for global bioequivalence for each comparator. If no k vector is provided, it will be set to the total number of endpoints on each comparator.}

\item{adjust}{alpha adjustment ( "k", "bon","sid","no","seq")}

\item{dropout}{vector with proportion of total population with dropout per arm}

\item{nsim}{number of simulated studies (default=5000)}

\item{lower}{initial value of N to be searched (default=2)}

\item{upper}{max value of N to be searched (default=500)}

\item{seed}{main seed}

\item{ncores}{number of cores used for the calculation if not provided it will use number of cores detected - 1}

\item{step.power}{initial step size is 2^step.power}

\item{step.up}{if TRUE steps up from 'lower', if FALSE steps down from 'upper'}

\item{pos.side}{if TRUE finds integer, i, closest to the root such that f(i) >0}

\item{maxiter}{maximum number of iterations}
}
\value{
An object simss that contains the following elements :
\describe{
\item{"response"}{ array with the sample sizes for each arm and aproximated achieved power with confidence intervals}
\item{"table.iter"}{data frame with the estimated sample size for each arm and power calculated at each searching iteration}
\item{"table.test"}{data frame that collects the total information of the simulation at each iteration}
\item{"param.u"}{parameters provided by the user}
\item{"param"}{ parameters used for the sample size calculation; as param.u are checked and modified in case of any inconsistent or missing information provided}
\item{"param.d"}{ parameters of design}
}
}
\description{
Sample size calculation given a target power (using uniroot.integer modified)----
}
\examples{

mu_list <- list(SB2 = c(AUCinf = 38703, AUClast = 36862, Cmax = 127.0),
                EUREF = c(AUCinf = 39360, AUClast = 37022, Cmax = 126.2),
                USREF = c(AUCinf = 39270, AUClast = 37368, Cmax = 129.2))

sigma_list <- list(SB2 = c(AUCinf = 11114, AUClast = 9133, Cmax = 16.9),
                   EUREF = c(AUCinf = 12332, AUClast = 9398, Cmax = 17.9),
                   USREF = c(AUCinf = 10064, AUClast = 8332, Cmax = 18.8))

# Equivalent boundaries
lequi.tol <- c(AUCinf = 0.8, AUClast = 0.8, Cmax = 0.8)
uequi.tol <- c(AUCinf = 1.25, AUClast = 1.25, Cmax = 1.25)


# arms to be compared
list_comparator <- list(EMA = c("SB2", "EUREF"),
                        FDA = c("SB2", "USREF"))

# Endpoints to be compared
list_y_comparator <- list(EMA = c("AUCinf", "Cmax"),
                         FDA = c("AUClast", "Cmax"))

# Run the simulation
result <- calopt(power = 0.9, # target power
                alpha = 0.05,
                mu_list = mu_list,
                sigma_list = sigma_list,
                lequi.tol = lequi.tol,
                uequi.tol = uequi.tol,
                list_comparator = list_comparator,
                list_y_comparator = list_y_comparator,
                adjust = "no",
                dtype = "parallel",
                ctype = "ROM",
                vareq = FALSE,
                lognorm = TRUE,
                ncores = 1,
                nsim = 50,
                seed = 1234)

}
\references{
Mielke, J., Jones, B., Jilma, B., & König, F. (2018). Sample size for multiple hypothesis testing in biosimilar development. Statistics in Biopharmaceutical Research, 10(1), 39-49.

Berger, R. L., & Hsu, J. C. (1996). Bioequivalence trials, intersection-union tests and equivalence confidence sets. Statistical Science, 283-302.
}
