% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSize.R
\name{derive_allocation_rate}
\alias{derive_allocation_rate}
\title{Derive Treatment Allocation Rate (TAR)}
\usage{
derive_allocation_rate(TAR, n_arms, verbose = FALSE)
}
\arguments{
\item{TAR}{Optional numeric vector specifying the allocation rate for each treatment arm. If missing, a default equal allocation rate is assigned.}

\item{n_arms}{Integer specifying the number of treatment arms.}

\item{verbose}{Logical, if \code{TRUE}, displays messages about the status of \code{TAR} derivation or assignment.}
}
\value{
A list representing the treatment allocation rate for each arm.
}
\description{
This function checks if \code{TAR} (treatment allocation rate) is provided. If \code{TAR} is missing, it assigns a default
equal allocation rate across all arms. It then converts \code{TAR} to a list format for further use.
Informational messages are displayed if \code{verbose} is set to \code{TRUE}.
}
\author{
Thomas Debray \email{tdebray@fromdatatowisdom.com}
}
