% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSize.R
\name{derive_endpoint_names}
\alias{derive_endpoint_names}
\title{Derive Endpoint Names}
\usage{
derive_endpoint_names(ynames_list, mu_list, verbose = FALSE)
}
\arguments{
\item{ynames_list}{Optional list of vectors with endpoint names for each arm.}

\item{mu_list}{Named list of means per treatment arm, where names can be used as endpoint names.}

\item{verbose}{Logical, if \code{TRUE}, displays messages about the derivation process.}
}
\value{
A list of endpoint names for each arm.
}
\description{
Derive Endpoint Names
}
\author{
Thomas Debray \email{tdebray@fromdatatowisdom.com}

This function derives endpoint names (\code{ynames_list}) from \code{mu_list} if \code{ynames_list}
is missing. If \code{ynames_list} is already provided, it confirms the names to the user when
\code{verbose} is set to \code{TRUE}.
}
