% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSize.R
\name{derive_TAR}
\alias{derive_TAR}
\title{Derive Treatment Allocation Rate (TAR)}
\usage{
derive_TAR(TAR = NULL, n_arms)
}
\arguments{
\item{TAR}{Numeric vector. Treatment allocation rates for each arm.}

\item{n_arms}{Numeric. The number of treatment arms that \code{TAR} must correspond to.}
}
\value{
A numeric vector of length \code{n_arms}, where any missing or NA values in \code{TAR}
have been replaced with ones.
}
\description{
This function validates and adjusts the treatment allocation rate (\code{TAR}) to ensure it matches
the number of treatment arms (\code{n_arms}). If \code{TAR} is missing or NULL, it is set to a default
vector of ones. If \code{TAR} is shorter than \code{n_arms}, missing values are replaced with ones.
If \code{TAR} contains NA values, they are replaced with ones. If \code{TAR} exceeds \code{n_arms} in length
or contains non-positive values, an error is raised.
}
\details{
\itemize{
\item If \code{TAR} is NULL or missing, it is set to a vector of ones of length \code{n_arms}.
\item If \code{TAR} is shorter than \code{n_arms}, missing values are replaced with ones and a warning is issued.
\item If \code{TAR} is longer than \code{n_arms}, an error is raised.
\item If \code{TAR} contains NA values, they are replaced with ones and a warning is issued.
\item If \code{TAR} contains negative or zero values, an error is raised.
}
}
\author{
Thomas Debray \email{tdebray@fromdatatowisdom.com}
}
