% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{info_msg}
\alias{info_msg}
\title{Helper function for conditional messages}
\usage{
info_msg(message, verbose)
}
\arguments{
\item{message}{A character string containing the message to display.}

\item{verbose}{Logical, if \code{TRUE}, the message is displayed; if \code{FALSE}, the message is suppressed.}
}
\value{
NULL (invisible). This function is used for side effects (displaying messages).
}
\description{
This function displays a message if the \code{verbose} parameter is set to \code{TRUE}.
It is useful for providing optional feedback to users during function execution.
}
\author{
Thomas Debray \email{tdebray@fromdatatowisdom.com}
}
\keyword{internal}
