% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Estimate sample size for phase I trial via simulation interactively}
\usage{
launch_app()
}
\description{
This function launch a shiny-based web app that allows the user calculate sample size for multiple arm multiple endpoint for phase I trial
The users are allowed to customize the parameters for simulation
}
\examples{
\dontrun{
   launch_app()

}

}
