% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mcsapply}
\alias{mcsapply}
\title{mcsapply}
\usage{
mcsapply(X, FUN, ..., simplify = TRUE, USE.NAMES = TRUE)
}
\arguments{
\item{X}{vector of iterations}

\item{FUN}{function}

\item{...}{additional parameters to pass}

\item{simplify}{simplify array}

\item{USE.NAMES}{use names in array}
}
\value{
vector output
}
\description{
An mc-version of the sapply function. https://stackoverflow.com/questions/31050556/parallel-version-of-sapply
}
\keyword{internal}
