% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_one_ep.r
\name{N_Mielke}
\alias{N_Mielke}
\title{N_Mielke}
\usage{
N_Mielke(
  power,
  Nmax,
  m,
  k,
  rho,
  sigma,
  true.diff,
  equi.tol,
  design,
  alpha,
  adjust = "no",
  seed = NULL,
  nsim = 10000
)
}
\arguments{
\item{power}{Desired power}

\item{Nmax}{Maximal sample size}

\item{m}{Number of endpoints (integer)}

\item{k}{Number of endpoints that must be successful (integer)}

\item{rho}{Correlation (this program assumes a constant correlation between the endpoints)}

\item{sigma}{Standard deviation of the endpoints (can be a vector of length m or a single value. If it is a single value, it is assumed that the standard deviation is constant for all endpoints). In case of the 2x2 crossover design, it is assumed that the input is the within-subject variance, in case of a parallel groups design, it is the standard deviation of the endpoint in the treatment group and it is assumed that this is identical for test and reference}

\item{true.diff}{Assumed true difference between test and reference (can be a vector of length m or a single value, see above)}

\item{equi.tol}{Equivalence margins, positive number, the interval is (-equi.tol, +equi.tol)}

\item{design}{Study design ("22co" for a 2x2 crossover design and "parallel" for a parallel groups design)}

\item{alpha}{Significance level}

\item{adjust}{Should an adjustment for multiplicity be applied ("no": no adjustment. "bon": Bonferroni, "k": k-adjustment)}

\item{seed}{Random Seed}

\item{nsim}{Number of iterations for power calculations (default: 10000)}
}
\value{
A vector with the achieved power (first entry) and the required sample size per sequence (second entry)
}
\description{
Sample size estimation for Multiple Hypothesis Testing (Mielke)
}
