% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{power_cal}
\alias{power_cal}
\title{power_cal}
\usage{
power_cal(n, nsim, param, param.d, seed, ncores)
}
\arguments{
\item{n}{sample size}

\item{nsim}{number of simulated studies}

\item{param}{list of parameters (mean,sd,tar)}

\item{param.d}{design parameters}

\item{seed}{main seed}

\item{ncores}{number of cores}
}
\value{
power calculated from a global list of comparators
}
\description{
Internal function to calculate the power across all comparators
}
\keyword{internal}
