% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_joint.R
\name{prepare_joint_count_inputs}
\alias{prepare_joint_count_inputs}
\title{Prepare inputs for a joint multi-arm count simulation}
\usage{
prepare_joint_count_inputs(
  rates,
  comparisons,
  exposure,
  margin_lower,
  margin_upper,
  alpha,
  endpoint_corr,
  list_margin_lower = NULL,
  list_margin_upper = NULL,
  dispersion = 0.1
)
}
\description{
Prepare inputs for a joint multi-arm count simulation
}
\keyword{internal}
