% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_methods.R
\name{print.countpower}
\alias{print.countpower}
\title{Print count-outcome power results}
\usage{
\method{print}{countpower}(x, ...)
}
\arguments{
\item{x}{A \code{countpower} object.}

\item{...}{Unused additional arguments.}
}
\description{
Print count-outcome power results
}
