% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_methods.R
\name{print.countss}
\alias{print.countss}
\title{Print count-outcome sample-size results}
\usage{
\method{print}{countss}(x, ...)
}
\arguments{
\item{x}{A \code{countss} object.}

\item{...}{Unused additional arguments.}
}
\description{
Print count-outcome sample-size results
}
