% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{print.simpower}
\alias{print.simpower}
\title{Print fixed-sample-size power results}
\usage{
\method{print}{simpower}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link[=simPower]{simPower()}}.}

\item{...}{Unused additional arguments.}
}
\value{
The input object, invisibly.
}
\description{
Print fixed-sample-size power results
}
