% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.simss_mielke.R
\name{print.simss_mielke}
\alias{print.simss_mielke}
\title{Print a Mielke sample-size result}
\usage{
\method{print}{simss_mielke}(x, ...)
}
\arguments{
\item{x}{A \code{simss_mielke} object.}

\item{...}{Unused additional arguments.}
}
\description{
Print a Mielke sample-size result
}
