% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ptv}
\alias{ptv}
\title{Compute p-values for a t-distribution with Fixed Degrees of Freedom}
\usage{
ptv(x, df, lower)
}
\arguments{
\item{x}{A numeric matrix (or vector) representing the random variables.}

\item{df}{A double specifying the degrees of freedom.}

\item{lower}{A logical value indicating whether to compute the lower-tail probability (\code{P(T <= x)}). If \code{FALSE}, the function returns the upper-tail probability (\code{P(T > x)}).}
}
\value{
A numeric matrix containing the computed cumulative distribution function (CDF) values (p-values).
}
\description{
Computes p-values for a given set of random variables under a t-distribution with fixed degrees of freedom.
}
\keyword{internal}
