% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sim_data}
\alias{sim_data}
\title{sim_data}
\usage{
sim_data(
  n,
  mu.arithmetic,
  mu.geometric = NULL,
  Sigma,
  CV = NULL,
  seed,
  dist = "normal"
)
}
\arguments{
\item{n}{sample size}

\item{mu.arithmetic}{arithmetic mean of endpoints}

\item{mu.geometric}{geometric mean of endpoints}

\item{Sigma}{variance covariance on what?? raw or transformed data}

\item{CV}{Coefficient of variation of what??}

\item{seed}{seed}

\item{dist}{assumed distribution : "normal" or "lognormal"}
}
\value{
Simulated endpoints for parallel design
}
\description{
Generate data for parallel design
}
\keyword{internal}
