% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_par_rom}
\alias{test_par_rom}
\title{test_par_rom_cpp}
\usage{
test_par_rom(
  n,
  muT,
  muR,
  SigmaT,
  SigmaR,
  lequi_tol,
  uequi_tol,
  alpha,
  dropout,
  typey,
  adseq,
  k,
  arm_seedT,
  arm_seedR,
  TART,
  TARR,
  vareq
)
}
\arguments{
\item{n}{integer number of subjects per arm}

\item{muT}{vector mean of endpoints on treatment arm}

\item{muR}{vector mean of endpoints on reference arm}

\item{SigmaT}{matrix covar-variance matrix on treatment arm across endpoints}

\item{SigmaR}{matrix covar-variance matrix on reference arm across endpoints}

\item{lequi_tol}{vector  lower equivalence tolerance band across endpoints}

\item{uequi_tol}{vector  upper equivalence tolerance band across endpoints}

\item{alpha}{vector alpha value across endpoints}

\item{dropout}{vector of size 2 with dropout proportion per arm (T,R)}

\item{typey}{vector with positions of primary endpoints}

\item{adseq}{boolean is used a sequential adjustment?}

\item{k}{integer minimum number of equivalent endpoints}

\item{arm_seedT}{integer seed for the simulation on treatment arm}

\item{arm_seedR}{integer seed for the simulation on reference arm}

\item{TART}{double treatment allocation rate for the treatment arm}

\item{TARR}{double treatment allocation rate for the reference arm}

\item{vareq}{boolean assumed equivalence variance between arms for the t-test}
}
\value{
mat(vector) with ptost and other simulated statistics such as mean (mu) and standard deviation(std) per sequence (0,1)-endpoint
}
\description{
Description : to simulate a parallel design and return the p-tost for ratio of means (ROM)
}
