% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_sample_size_limits}
\alias{validate_sample_size_limits}
\title{Check Sample Size Limits}
\usage{
validate_sample_size_limits(lower, upper)
}
\arguments{
\item{lower}{Numeric. The initial lower limit for the search range.}

\item{upper}{Numeric. The initial upper limit for the search range.}
}
\value{
NULL. If the checks pass, the function returns nothing. If the checks fail, it stops execution with an error message.
}
\description{
Validates that the upper and lower limits are numeric and that the upper limit is greater than the lower limit.
}
\author{
Thomas Debray \email{tdebray@fromdatatowisdom.com}
}
\keyword{internal}
