% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_tar}
\alias{validate_tar}
\title{Validate Treatment Allocation Rate (TAR)}
\usage{
validate_tar(TAR = NULL, n_arms)
}
\arguments{
\item{TAR}{Numeric vector. Treatment allocation rates for each arm.}

\item{n_arms}{Numeric. The number of treatment arms that \code{TAR} must correspond to.}
}
\value{
NULL (used for validation only).
}
\description{
This function validates whether the length of the treatment allocation rate (\code{TAR}) matches
the number of treatment arms (\code{arm_names}). If the lengths do not match, it throws an error.
}
\author{
Thomas Debray \email{tdebray@fromdatatowisdom.com}
}
