% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.simss.R
\name{confint.simss}
\alias{confint.simss}
\title{Confidence Interval for Achieved Power from simss object}
\usage{
\method{confint}{simss}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"simss"} returned by a sampleSize function}

\item{...}{Additional arguments (currently unused).}
}
\value{
A named numeric vector with two elements:
\describe{
\item{Achieved Power}{Achieved power.}
\item{Lower}{Lower bound of the confidence interval.}
\item{Upper}{Upper bound of the confidence interval.}
}
}
\description{
Confidence Interval for Achieved Power from simss object
}
\examples{
# Assume `res` is a result from `sampleSize()`
# confint(res)

}
