% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.normalize_internal_distribution}
\alias{.normalize_internal_distribution}
\title{Calculate the power across all comparators}
\usage{
.normalize_internal_distribution(param.d)
}
\arguments{
\item{param.d}{design parameters}
}
\value{
power calculated from a global list of comparators
}
\description{
Internal function to calculate the power across all comparators
}
\keyword{internal}
