% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_par.R
\name{get_par}
\alias{get_par}
\title{get_par}
\usage{
get_par(
  mu_list,
  varcov_list,
  TAR_list,
  type_y = NA,
  arm_names = NA,
  y_names = NA
)
}
\arguments{
\item{mu_list}{list of mu vectors, each element corresponds to a comparator with a mu vector of size number of endpoints.}

\item{varcov_list}{list of var-cov matrices, each element corresponds to a comparator with a varcov matrix of size number of endpoints X number of endpoints.}

\item{TAR_list}{list of allocation rates, each element corresponds to a comparator with a allocation number(integer or not).}

\item{type_y}{vector with the type of enpoints: primarly endpoint(1), otherwise (2).}

\item{arm_names}{vector with the names of the arms.}

\item{y_names}{vector with the names of the endpoints.}
}
\value{
named list with arm_names of mu, varcov and TAR to been passed to our function
}
\description{
Function to pass the parameters of endpoints and comparators in a list form required as input in the calopt function
}
