% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sims.R
\name{plot.simss}
\alias{plot.simss}
\title{plot.simss}
\usage{
\method{plot}{simss}(x, ...)
}
\arguments{
\item{x}{Object type simss}

\item{\ldots}{Additional arguments which are passed to simss object.}
}
\value{
Plot of power vs total sample size per comparator and total of comparators
}
\description{
Plot of power vs sample size for all comparators and total of comparators
}
