% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSize_Mielke.R
\name{power_dom}
\alias{power_dom}
\title{Power Calculation for Hypothesis Testing of Difference of Means (DOM)}
\usage{
power_dom(
  seed,
  mu_test,
  mu_control,
  sigma_test,
  sigma_control,
  N_test,
  N_control,
  lb,
  ub,
  alpha = 0.05,
  nsim = 10000
)
}
\arguments{
\item{seed}{Integer. A seed value for reproducibility of the simulations.}

\item{mu_test}{Numeric. The arithmetic mean of the test group.}

\item{mu_control}{Numeric. The arithmetic mean of the control group.}

\item{sigma_test}{Numeric. The standard deviation of the test group.}

\item{sigma_control}{Numeric. The standard deviation of the control group.}

\item{N_test}{Integer. The sample size of the test group.}

\item{N_control}{Integer. The sample size of the control group.}

\item{alpha}{Numeric. The significance level for the hypothesis test. Default is 0.05.}

\item{nsim}{Integer. The number of simulations to perform. Default is 10,000.}
}
\value{
Numeric. The estimated power based on the simulations.
}
\description{
This function calculates the power of hypothesis testing for the difference of means (DOM) between two groups, using simulations to estimate the achieved power based on the provided parameters such as sample sizes, means, standard deviations, and significance level.
}
\keyword{internal}
