% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{print.simss}
\alias{print.simss}
\title{Display the summary results of the sample size estimation}
\usage{
\method{print}{simss}(x, ...)
}
\arguments{
\item{x}{An object of class "simss"}

\item{...}{Optional arguments to be passed from or to other methods.}
}
\description{
Display the summary results of the sample size estimation
}
\author{
Thomas Debray \email{thomas.debray@gmail.com}
}
