% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ptv}
\alias{ptv}
\title{Compute p-values with Fixed Degrees of Freedom}
\usage{
ptv(x, df, lower)
}
\arguments{
\item{x}{A numeric matrix (or vector) of random variables.}

\item{df}{A double specifying the degrees of freedom.}

\item{lower}{A boolean indicating whether to calculate the lower-tail probability
(i.e., P(T <= x)). If \code{false}, computes the upper-tail probability (P(T > x)).}
}
\value{
A numeric matrix containing the computed CDF values (p-values).
}
\description{
Computes the cumulative distribution function (CDF) values (p-values)
for a given set of random variables, assuming a t-distribution with fixed
degrees of freedom.
}
