% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ptv}
\alias{ptv}
\title{ptv}
\usage{
ptv(x, df, lower)
}
\arguments{
\item{x}{mat(vector) random variable}

\item{df}{double degrees of freedom}

\item{lower}{boolean, calculate the probability of the region where the random variable is less than or equal to x}
}
\value{
mat(vector) with the cumulative distribution function (p-values)
}
\description{
Description : to get the p-value with fixed degrees of freedom across the endpoints
}
