% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_one_ep.r
\name{sign_Mielke}
\alias{sign_Mielke}
\title{Simulated Test Statistic for Given Settings Using Mielke's Method}
\usage{
sign_Mielke(
  N,
  m,
  k,
  R,
  sigma,
  true.diff,
  equi.tol = log(1.25),
  design,
  alpha = 0.05,
  adjust = "no"
)
}
\arguments{
\item{N}{Integer. The number of subjects per sequence.}

\item{m}{Integer. The number of endpoints.}

\item{k}{Integer. The number of endpoints that must be successful to consider the test a success.}

\item{R}{Matrix. The correlation matrix between the endpoints (e.g., generated with the \code{variance.const.corr()} function), with dimensions m x m.}

\item{sigma}{Numeric. The standard deviation of the endpoints. This can either be a vector of length m (one for each endpoint) or a single value. If a single value is provided, it is assumed that the standard deviation is constant across all endpoints. In the case of a 2x2 crossover design, the input is the within-subject variance; for a parallel group design, it represents the standard deviation in the treatment group, assumed identical for both test and reference groups.}

\item{true.diff}{Numeric. The assumed true difference between the test and reference. This can be a vector of length m (one for each endpoint) or a single value.}

\item{equi.tol}{Numeric. Equivalence margins, with the interval being (-equi.tol, +equi.tol).}

\item{design}{Character. The study design, either "22co" for a 2x2 crossover design or "parallel" for a parallel groups design.}

\item{alpha}{Numeric. The significance level.}

\item{adjust}{Character. The method for multiplicity adjustment. Options are "no" for no adjustment, "bon" for Bonferroni correction, or "k" for k-adjustment.}
}
\value{
A realization of the simulated test statistic for the given setting. If values are provided as a single number (constant for all endpoints), a vector is created.
}
\description{
This function simulates a test statistic for a given setting using Mielke's method, accounting for multiple endpoints, correlation structure, and study design.
}
\keyword{internal}
