% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimTOST-package.R
\docType{package}
\name{SimTOST}
\alias{SimTOST-package}
\alias{SimTOST}
\title{Sample Size Estimation via Simulation}
\description{
SimTOST: A Package for Sample Size Simulations
}
\details{
The SimTOST package provides tools for simulating sample sizes, calculating power,
and assessing type-I error for various statistical scenarios.
}
\references{
Mielke, J., Jones, B., Jilma, B. & König, F. Sample Size for Multiple Hypothesis Testing in Biosimilar Development. \emph{Statistics in Biopharmaceutical Research} 10, 39–49 (2018).
}
\seealso{
\link{calopt}
}
\author{
\strong{Maintainer}: Thomas Debray \email{tdebray@fromdatatowisdom.com} (\href{https://orcid.org/0000-0002-1790-2719}{ORCID})

Authors:
\itemize{
  \item Johanna Muñoz \email{johanna.munoz@fromdatatowisdom.com}
  \item Dewi Amaliah \email{dewi.amaliah@fromdatatowisdom.com}
}

Other contributors:
\itemize{
  \item Wei Wei \email{wei.wei@biogen.com} [contributor]
  \item Marian Mitroiu \email{marian.mitroiu@biogen.com} [contributor]
  \item Biogen Inc [copyright holder, funder]
}

}
