% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_par_rom}
\alias{test_par_rom}
\title{Simulate a Parallel Design and Compute Ratio of Means (ROM)}
\usage{
test_par_rom(
  n,
  muT,
  muR,
  SigmaT,
  SigmaR,
  lequi_tol,
  uequi_tol,
  alpha,
  dropout,
  typey,
  adseq,
  k,
  arm_seedT,
  arm_seedR,
  TART,
  TARR,
  vareq
)
}
\arguments{
\item{n}{integer number of subjects per arm}

\item{muT}{vector mean of endpoints on treatment arm}

\item{muR}{vector mean of endpoints on reference arm}

\item{SigmaT}{matrix covar-variance matrix on treatment arm across endpoints}

\item{SigmaR}{matrix covar-variance matrix on reference arm across endpoints}

\item{lequi_tol}{vector  lower equivalence tolerance band across endpoints}

\item{uequi_tol}{vector  upper equivalence tolerance band across endpoints}

\item{alpha}{vector alpha value across endpoints}

\item{dropout}{vector of size 2 with dropout proportion per arm (T,R)}

\item{typey}{vector with positions of primary endpoints}

\item{adseq}{boolean is used a sequential adjustment?}

\item{k}{integer minimum number of equivalent endpoints}

\item{arm_seedT}{integer seed for the simulation on treatment arm}

\item{arm_seedR}{integer seed for the simulation on reference arm}

\item{TART}{double treatment allocation rate for the treatment arm}

\item{TARR}{double treatment allocation rate for the reference arm}

\item{vareq}{Boolean. If \code{TRUE}, assumes equal variance between arms and applies
Schuirmann’s two one-sided tests (TOST) for equivalence using a pooled variance.}
}
\value{
A numeric matrix containing the following simulated statistics:
\itemize{
\item \code{ptost}: p-values for equivalence testing.
\item \code{mu0}, \code{mu1}: Mean values for the reference and treatment groups.
\item \code{sd0}, \code{sd1}: Standard deviations for the reference and treatment groups.
\item \code{totaly}: Indicator (1/0) for equivalence success.
}
}
\description{
Simulates a parallel-group design and calculates the p-value for the
ratio of means (ROM) using an equivalence test.
}
\details{
\itemize{
\item The function simulates a parallel-group study design and applies an equivalence test
using the ratio of means (ROM) approach.
\item Adjusts for dropout rates (\code{dropout}) and treatment allocation rates (\code{TART}, \code{TARR}).
\item Computes test statistics and determines whether the number of significant endpoints meets \code{k}.
\item Uses sequential testing if \code{adseq = TRUE}.
\item If \code{vareq = TRUE}, assumes equal variance between treatment and reference groups,
applying Schuirmann’s two one-sided tests (TOST) for equivalence.
}
}
