% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uniroot.integer.mod}
\alias{uniroot.integer.mod}
\title{uniroot.integer.mod}
\usage{
uniroot.integer.mod(
  f,
  power,
  lower = lower,
  upper = upper,
  step.power = step.power,
  step.up = step.up,
  pos.side = pos.side,
  maxiter = maxiter,
  ...
)
}
\arguments{
\item{f}{function for which a root is needed}

\item{power}{target power value}

\item{lower}{minimum allowable root}

\item{upper}{maximum allowable root}

\item{step.power}{initial step size is 2^step.power}

\item{step.up}{if TRUE steps up from 'lower', if FALSE steps down from 'upper'}

\item{pos.side}{if TRUE finds integer, i, closest to the root such that f(i) >0}

\item{maxiter}{maximum number of iterations}

\item{...}{additional arguments to 'f'.}
}
\value{
a list with the following elements:
root= the integer on the correct side of the root
f.root	= value of f at root  (output= main output with final power, output.test= output table with test at endpoint level)
iter	= number of times f was evaluated
table.iter = data.frame with the estimated N and power at each iteration
table.test = data.frame with the test at endpoint level for a given N and each simulation draw
}
\description{
Optimizer Uniroot integer modified from the ssanv function package https://github.com/cran/ssanv/blob/master/R/uniroot.integer.R
}
\keyword{internal}
